// lib/api/generated/models/models-brand-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Huerray Backend
 * Comprehensive API for connecting content creators with brands for marketing campaigns and gig opportunities
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thrypes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ModelsContentMedia } from './models-content-media';
// May contain unused imports in some cases
// @ts-ignore
import type { ModelsUserResponse } from './models-user-response';
// May contain unused imports in some cases
// @ts-ignore
import type { UtilsCompanySize } from './utils-company-size';
// May contain unused imports in some cases
// @ts-ignore
import type { UtilsCountryCode } from './utils-country-code';

export interface ModelsBrandResponse {
    'brand_id'?: string;
    'brand_status'?: string;
    'building_number'?: string;
    'category'?: string;
    'city'?: string;
    'company_description'?: string;
    'company_name'?: string;
    'company_size'?: UtilsCompanySize;
    'country'?: UtilsCountryCode;
    'created_at'?: string;
    'id'?: string;
    'postal_code'?: string;
    'preferred_contact_email'?: string;
    'preferred_contact_phone'?: string;
    'profile_photo'?: ModelsContentMedia;
    'registration_number'?: string;
    'state'?: string;
    'status_change_comments'?: string;
    'street'?: string;
    'updated_at'?: string;
    'user'?: ModelsUserResponse;
    'user_id'?: string;
    'vat_id'?: string;
    'website_url'?: string;
}